+++ app/controllers/shortener_controller.rb
@@ -1,6 +1,0 @@
-class ShortenerController < ApplicationController
-
-  def new
-  end
-
-end