+++ app/controllers/shorteners_controller.rb
@@ -1,0 +1,8 @@
+class ShortenersController < ApplicationController
+  def new
+  end
+
+  def create
+    render plain: params[:shortener].inspect
+  end
+end